fix(ci): consolidate invocation wait files and reconcile unit latency inventory
Backend Lint pkg-file-count holds pkg/services/factory_sessions/internal/
sessionservice at 38 files, so the event-driven invocation waiter moves into
runtime_invocation.go and its tests join work_runtime_resolver_test.go with no
behavior change.

Backend Unit Latency pins an exact reference test inventory; the seven unit
tests this branch adds are now recorded in the baseline inventory and the
reviewed count moves 18239 -> 18246, matching the established reconcile
pattern (bf5a6e33b7).
