+++ docs/internal/baselines/go-unit-lane-latency-budget.v1.json
@@ -6161,6 +6161,7 @@
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_MapsTimeoutAndCancellation",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_MapsTimeoutAndCancellation/cancellation",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_MapsTimeoutAndCancellation/timeout",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_NilSessionWaiterFallsBackToWaitNext",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_PreservesTerminalFailureClassifications",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_PreservesTerminalFailureClassifications/blocked",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_PreservesTerminalFailureClassifications/failed",
@@ -6171,6 +6172,8 @@
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_ReturnsFailedFromStoppedScopedFailure",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_ReturnsFailedWhileActiveWorkHasScopedFailure",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_ReturnsSubmittedTerminalContent",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_SessionWaiterWithNilReleaseCompletes",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwnerWait_WaitSessionWaiterIsPreferredAndReleasedOnce",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwner_CompatibilityContentRoutesToPrimarySignatureArgument",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwner_MapsCancellationArrivingDuringSubmissionToInvocationOutcome",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation::TestSessionOwner_PreparedCLIArgumentsRetainCanonicalValuesOrderAndProvenance",
@@ -7066,12 +7069,16 @@
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestDetachedRouterRoutesSessionOperationsToOwningGateway/second",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestDurableCallSitesRouteThroughOwnerCapabilityNotHostAccessor",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestDurableCapabilityDoesNotTouchCollidingLiveSession",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestEventDrivenInvocationWaiterHonorsContextCancellation",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestEventDrivenInvocationWaiterWakesOnRelevantEventBeforeFallback",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestInvocationOutcomeRelevantEventSelectsStateAndLifecycleTypes",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestLiveControlCapability_CanceledClosePreservesLiveSession",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestLiveControlCapability_CompletesLifecycleAndRetiresCanonicalSession",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestLiveControlCapability_OpenListReadPreservesCanonicalIdentity",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestLiveControlCapability_OpenPauseResumePreservesLifecycleResults",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestLiveControlCapability_PreservesTypedRejectionAndCancellation",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestNewWithResponseService_ComposesLiveRuntimeWithoutRegistryEffects",
+      "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestRelayInvocationWakeEventsCoalescesBurstsWithoutBlocking",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestResolveWorkRuntimeDiscardsProjectionFromReplacedGeneration",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestResolveWorkRuntimeReleasesProjectionAfterSessionDisappears",
       "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/sessionservice::TestRetireWorkAdmissionProjectionAfterRepeatedReplacement",
